+++ splitio_platform_interface/lib/split_certificate_pinning_configuration.dart
@@ -1,9 +1,11 @@
+/// Certificate pinning configuration.
 class CertificatePinningConfiguration {
-
   final Map<String, List<String>> _pins = {};
 
+  /// Returns the pins.
   Map<String, List<String>> get pins => _pins;
 
+  /// Adds a pin for the given host.
   CertificatePinningConfiguration addPin(String host, String pin) {
     pin = pin.trim();
     if (pin.isEmpty) {