+++ splitio_platform_interface/lib/split_prerequisite.dart
@@ -1,15 +1,20 @@
+/// Prerequisite class.
 class Prerequisite {
   final String _name;
 
   final Set<String> _treatments;
 
+  /// The feature flag name of the prerequisite.
   String get name => _name;
 
+  /// The treatments of the prerequisite.
   Set<String> get treatments => _treatments;
 
+  /// Creates a new Prerequisite instance.
   Prerequisite(this._name, this._treatments);
 
-  static Prerequisite fromEntry(el) {
+  /// Creates a Prerequisite instance from a map.
+  static Prerequisite fromEntry(Map<String, dynamic> el) {
     final String name = (el['n'] ?? el['n:'] ?? '').toString();
     final List<dynamic> rawTreatments = (el['t'] as List<dynamic>?) ?? [];
     final Set<String> treatments =
@@ -26,7 +31,8 @@ class Prerequisite {
     }''';
   }
 
-  equals(Prerequisite other) {
+  /// Checks if this Prerequisite is equal to another Prerequisite.
+  bool equals(Prerequisite other) {
     return name == other.name && treatments == other.treatments;
   }