+++ splitio_web/CHANGELOG.md
@@ -1,4 +1,15 @@
-# 1.0.0 (January 16, 2026)
-- Initial release. Web implementation of `splitio` based on Split Browser SDK v1.6.0.
+# 1.1.1 (Jan 28, 2026)
+* Bug fix: corrected asset path for loading the Browser SDK.
 
-# 1.0.0-rc.1 (January 15, 2026)
+# 1.1.1-rc.1 (Jan 28, 2026)
+
+# 1.1.0 (Jan 28, 2026)
+* Added new configuration for Fallback Treatments, which allows setting a treatment value and optional config to be returned in place of "control", either globally or by flag. Read more in our docs.
+* Updated Browser SDK to `1.6.1`.
+
+# 1.1.0-rc.1 (Jan 28, 2026)
+
+# 1.0.0 (Jan 16, 2026)
+* Initial release. Web implementation of `splitio` based on Split Browser SDK `1.6.0`.
+
+# 1.0.0-rc.1 (Jan 15, 2026)